Appからビジネスロジックを分離し、DriveListManagerを作成
