+++ App.tsx
@@ -1,9 +1,10 @@
 import * as React from 'react';
-import { Text, View, Button, StyleSheet, FlatList, Alert } from 'react-native';
-import Constants from 'expo-constants';
+import { View, Button, StyleSheet, FlatList } from 'react-native';
 import Modal from "react-native-modal";
 
+import {Drive} from './components/Drive';
 import DriveList, {DeviceListProps} from './components/DriveList';
+import DriveListManager from './components/DriveListManager';
 import PointNameDialog, { PointNameDialogState } from './components/PointNameDialog';
 
 interface AppState {
@@ -11,38 +12,24 @@ interface AppState {
   isModalVisible: boolean;
 }
 
-interface Drive {
-  id: number;
-  pointName?: string;
-  arrivalTime?: string;
-  departureTime?: string;
-}
-
-class DriveImpl implements Drive {
-  id: number;
-  pointName?: string;
-  arrivalTime?: string;
-  departureTime?: string;
-
-  constructor(id: number) {
-    this.id = id;
-  }
-}
-
-
 /**
  * ApplicationComponent
  */
 export default class App extends React.Component<{}, AppState> {
+
+  drivesManager: DriveListManager = new DriveListManager(this);
 
   /**
    * Constructor
    */
   constructor(props: {}) {
     super(props);
-    this.state = this.initializeState();
+    this.state = {
+      drives: this.drivesManager.drives,
+      isModalVisible: false
+    } as AppState;
   }
-
+ 
   /**
    * render main view.
    */
@@ -81,45 +68,8 @@ export default class App extends React.Component<{}, AppState> {
     );
   };
 
-  /**
-   * handle Record button touched.
-   */
   handleRecordBtnClick = () => {
-    let updatedDrive;
-    let addDrive = false;
-    let tmpDrives = [...this.state.drives]
-                      .map((drive, index) => {
-                        if (index !== this.state.drives.length - 1) return drive;
-                        updatedDrive = this.updateDriveInfo(drive);
-                        if (updatedDrive === null) {
-                          updatedDrive = {...drive};
-                          addDrive = true;
-                        }
-                        return updatedDrive;
-                      });
-    if (addDrive) {
-      // 最後の地点の出発時刻まで記録済なので、新しい地点オブジェクトを追加
-      const newDrive = new DriveImpl(this.state.drives.length);
-      newDrive.arrivalTime = (new Date()).toLocaleTimeString();
-      tmpDrives.push(newDrive);
-    }
-    this.setState({drives: tmpDrives} as AppState);
-  }
-
-  private updateDriveInfo = (drive: Drive) => {
-    const newDrive = {...drive};
-    if (drive.arrivalTime === undefined) {
-      // このコードを通ることはないはず
-      newDrive.arrivalTime = (new Date()).toLocaleTimeString();
-    } else if (drive.pointName === undefined) {
-      this.toggleModal(true);
-    } else if (drive.departureTime === undefined) {
-      newDrive.departureTime = (new Date()).toLocaleTimeString();
-    } else {
-      // 最後の地点の出発時刻まで記録済
-      return null;
-    }
-    return newDrive;
+    this.drivesManager.addNewRecord();
   }
 
   /**
@@ -127,49 +77,13 @@ export default class App extends React.Component<{}, AppState> {
    */
   handleDialogDismiss = (value: PointNameDialogState) => {
     if (value !== undefined) {
-      let tmpDrives = [...this.state.drives]
-      .map((drive, index) => {
-        if (index !== this.state.drives.length - 1) return drive;
-        const newDrive = {...drive};
-        newDrive.pointName = value.pointName;
-        return newDrive;
-      });
-      this.setState({drives: tmpDrives} as AppState)
+      this.drivesManager.addPointName(value.pointName);
     }
     this.toggleModal(false);
   }
 
   toggleModal = (v: boolean) => {
     this.setState({isModalVisible: v} as AppState);
-  }
-
-  /**
-   * initialize state value;
-   */
-  private initializeState = () => {
-    return {
-      drives: [
-        {
-          id: 0,
-          arrivalTime: "",
-          pointName: 'スタート地点',
-          departureTime: ""
-        },
-        {
-          id: 1,
-          arrivalTime: "",
-          pointName: '村営駐車場',
-          departureTime: ""
-        },
-        {
-          id: 2,
-          arrivalTime: "",
-          pointName: '休憩所',
-          departureTime: ""
-        },
-      ],
-      isModalVisible: false
-    } as AppState
   }
 
 }
